SLING-8056 : Add support for automatically updating versions in a feature
